requirements.txt
pipecat-ai[openai, google, silero]
python-dotenv
pyaudio
sounddevice
scipy